Expliticly check for git-lfs for gA
External gA data package relies on git-lfs to pull/download its data
files. If this is not present, the build/install will proceed, but the
files will only hold the LFS pointers. This leads to tests failing
and/or a broken install.

When using gA, explicitly check for git-lfs and error out if it is
not present.