Improve thread safety og static resources
As reported in #28, bxdecay0 can produce segfaults when running as
the primary generator of a multithreaded Geant4 application. Triaged
in this use case to the use of statics in the pattern:

```
some_t function() {
  static thing_t thing;
  if (thing.not_init()) thing.fill();
  ...
  return thing;
}
```

Whilst C++11 guarantees the initialization of `thing` to be thread
safe (at least for standard types), the code following the init
statement is not. For example, a vector could be initialzed then
filled if empty, but multiple threads could contend over the scope
that does the filling.

Decouple static initialization by factoring the filling of the
static objects into separate functions. As the initialization
of the static is only called once, so is the initialization function
and thus removing thread contention issues.

Note that the BinReloc functions are not modified using this pattern,
but as the calls to these functions are only in bxdecay0::get_library_dir()
and its init function, these should be sufficiently protected. At worst
a lock_guard is needed in this one place.